Dockerfile now expects services from Kubernetes mount
AS we're trying to use the custom SP Integration support, we needed to allow CAs to create new service json files on the fly.  So far, we've been having Kubernetes mount a directory containing COB service definitions.  However, these mounts are read-only.  CAS would crash when attempting to write a new service json file to that directory.

This commit changes the Dockerfile to expect Kubernetes to mount the initial service definitions somewhere safe, then copies them to the official place in an entrypoint script.

Updates #14